HuggingFaceTransformers

Documentation for HuggingFaceTransformers.

HuggingFaceTransformers.HuggingFaceTransformers

HuggingFaceTransformers.Model

HuggingFaceTransformers.apply_chat_template

HuggingFaceTransformers.decode

HuggingFaceTransformers.encode

HuggingFaceTransformers.from_pretrained

HuggingFaceTransformers.from_pretrained

HuggingFaceTransformers.from_pretrained

HuggingFaceTransformers.generate

HuggingFaceTransformers.tensor

HuggingFaceTransformers.HuggingFaceTransformers — Module

HuggingFaceTokenizers

A Julia wrapper around HuggingFace's Tokenizers Python library.

See https://huggingface.co/docs/tokenizers/en/index for official documentation.

HuggingFaceTransformers.Model — Type

Tokenizer

A wrapper around a Python tokenizer.

HuggingFaceTransformers.apply_chat_template — Method

apply_chat_template(tokenizer::AutoTokenizer, messages::Vector{Dict{String, String}}, tokenize=true, add_generation_prompt=false, kwargs...)

Applies a chat template (if one is associated with this model's tokenizer). Keyword arguments include tokenize (bool) and add_generation_prompt (bool).

prompt = "Write me a poem."
messages = [
    Dict("role" => "system", "content" => "You are Bob, a helpful assistant."),
    Dict("role" => "user", "content" => prompt)]

text = apply_chat_template(tokenizer, messages, tokenize=false, add_generation_prompt=true)

HuggingFaceTransformers.decode — Method

decode(tokenizer::Tokenizer, ids::Vector{Int}, skip_special_tokens = false) -> String

Decode a sequence of token IDs back into text.

HuggingFaceTransformers.encode — Method

encode(tokenizer::Tokenizer, text::String, kwargs...) -> Vector{Int}

Encode a single text string into tokens and their corresponding IDs.

HuggingFaceTransformers.from_pretrained — Method

from_pretrained(model_path::String) -> Tokenizer

Load a pretrained AutoTokenizer. m can be a model name or a path to a local model, or a HuggingFace URL. Compared to HuggingFaceTokenizers.jl, this can use eg. chat templates.

HuggingFaceTransformers.from_pretrained — Method

from_pretrained(CausalLM, model_path::String) -> Model

Load a pretrained model. m can be a model name or a path to a local model, or a HuggingFace URL.

HuggingFaceTransformers.from_pretrained — Method

from_pretrained(Transformer, model_path::String) -> Model

Load a pretrained model. m can be a model name or a path to a local model, or a HuggingFace URL.

HuggingFaceTransformers.generate — Method

generate(model::CausalLM, prefix_tokens::Vector{Int64}; kw...) -> Vector{Array}
generate(model::CausalLM, prompt::String, tokenizer; kw...) -> String

Generate text from a model.

HuggingFaceTransformers.tensor — Method

tensor(x::Py) -> Array

Convert a PyTorch tensor to a Julia array, detaching the tensor.